+++ src/Messenger/Message/JournalLogMessage.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace AnzuSystems\CommonBundle\Messenger\Message;
 
-final class AppLogMessage extends AbstractLogMessage
+final class JournalLogMessage extends AbstractLogMessage
 {
 }